+++ test/running_modules/run_tests.sh
@@ -61,3 +61,35 @@ should_fail run --module module/no_main_function
 
 # Main function with wrong arity
 should_fail run --module module/wrong_arity
+should_fail run --module wrong_test_arity
+should_fail run --module wrong_dev_arity
+
+# Test modules
+should_succeed test --target erlang
+should_succeed test --target javascript --runtime nodejs
+should_succeed test --target javascript --runtime deno
+
+should_succeed run --module module_in_test --target erlang
+should_succeed run --module module_in_test --target javascript --runtime nodejs
+should_succeed run --module module_in_test --target javascript --runtime deno
+
+# Nested module in test
+should_succeed run --module nested/module_in_test --target erlang
+should_succeed run --module nested/module_in_test --target javascript --runtime nodejs
+should_succeed run --module nested/module_in_test --target javascript --runtime deno
+
+
+# Dev modules
+should_succeed dev --target erlang
+should_succeed dev --target javascript --runtime nodejs
+should_succeed dev --target javascript --runtime deno
+
+should_succeed run --module module_in_dev --target erlang
+should_succeed run --module module_in_dev --target javascript --runtime nodejs
+should_succeed run --module module_in_dev --target javascript --runtime deno
+
+# Nested module in dev
+should_succeed run --module nested/module_in_dev --target erlang
+should_succeed run --module nested/module_in_dev --target javascript --runtime nodejs
+should_succeed run --module nested/module_in_dev --target javascript --runtime deno
+